Refactor - VoteForm submission code to React standards
- Conditional render on submission
- Lift state up to App level

[Fixes: #9]
