+++ src/App.js
@@ -1,13 +1,48 @@
-import React from 'react';
+import React, { useState } from 'react';
 import './App.css';
 import VoteForm from './components/VoteForm'
-import VoteResult from './components/VoteResult'
 
 function App() {
+  const myChoiceApiVoteURL = 'https://mychoice-api.herokuapp.com/api/vote'
+  const [hasVoted, setHasVoted] = useState(false)
+  const [value, setValue] = useState("")
+
+  const handleChange = e => {
+    setValue(e.target.value)
+  }
+  
+  const handleSubmit = e => {
+    e.preventDefault();
+    let data = { candidate_name: value }
+    const myRequestInit = {
+      method: 'POST',
+      headers: { 
+        'Content-Type': 'application/json',
+        'Accept': 'application/json',
+      },
+      body: JSON.stringify(data)
+    }
+
+
+    fetch(myChoiceApiVoteURL, myRequestInit)
+      .then( response => response.json() )
+      .then( data => {
+        console.log('Vote submitted successfully: ', data);
+        setHasVoted(true)
+      })
+      .catch( error => {
+        console.log('Error submitting vote.. ', error)
+      })
+  }
+
+
+
   return (
     <div className="App">
-      <VoteForm />
-      <VoteResult />
+      <VoteForm
+        formChange={ handleChange }
+        formSubmit={ handleSubmit }
+        voted={ hasVoted }/>
     </div>
   );
 }
